refactoring: refactored dart main         новый файл:    lib/counter.dart         изменено:      lib/main.dart         новый файл:    lib/root_wg.dart
