+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_inherited_widget/root_wg.dart';
 
 void main() {
   runApp(MyApp());
@@ -33,8 +34,8 @@ class _MyHomePageState extends State<MyHomePage> {
 
   int get counterValue => _counter;
 
-  void _incrementCounter() => setState(() =>  _counter++);
-  void _decrementCounter() => setState(() => _counter--);
+  void incrementCounter() => setState(() =>  _counter++);
+  void decrementCounter() => setState(() => _counter--);
 
   @override
   Widget build(BuildContext context) {
@@ -52,65 +53,10 @@ class _MyHomePageState extends State<MyHomePage> {
         ],
       ),
       floatingActionButton: FloatingActionButton(
-        onPressed: _incrementCounter,
+        onPressed: incrementCounter,
         tooltip: 'Increment',
         child: Icon(Icons.add),
       ), // This trailing comma makes auto-formatting nicer for build methods.
-    );
-  }
-}
-
-class AppRootWidget extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    final rootWidgetState = MyInheritedWidget.of(context).myState;
-    return Card(
-      elevation: 4.0,
-      child: Column(
-        children: <Widget>[
-          Text('(Root Widget)', style: Theme.of(context).textTheme.display1),
-          Text('${rootWidgetState.counterValue}', style: Theme.of(context).textTheme.display1),
-          SizedBox(height: 50,),
-          Row(
-            mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-            children: <Widget>[
-              Counter(),
-              Counter(),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class Counter extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    final rootWidgetState = MyInheritedWidget.of(context).myState;
-    return Card(
-      margin: EdgeInsets.all(4.0).copyWith(bottom: 32),
-      color: Colors.yellowAccent,
-      child: Column(
-        children: <Widget>[
-          Text('(Children Widget)'),
-          Text('(${rootWidgetState.counterValue})', style: Theme.of(context).textTheme.display1),
-          ButtonBar(
-            children: <Widget>[
-              IconButton(
-                onPressed: () => rootWidgetState._decrementCounter(),
-                icon: Icon(Icons.remove),
-                color: Colors.red
-              ),
-              IconButton(
-                onPressed: () => rootWidgetState._incrementCounter(),
-                icon: Icon(Icons.add),
-                color: Colors.green,
-              ),
-            ],
-          )
-        ],
-      ),
     );
   }
 }
